+++ MyElectricalShop/MyElectricalShop.Application/ActionMethods/VoltageLevels/Delete/DeleteVoltageLevelRequest.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using MyElectricalShop.Domain.Interfaces;
+using Core.Exceptions;
 
 namespace MyElectricalShop.Application.ActionMethods.VoltageLevels.Delete
 {
@@ -25,6 +26,9 @@ namespace MyElectricalShop.Application.ActionMethods.VoltageLevels.Delete
         public async Task<Unit> Handle(DeleteVoltageLevelRequest request, CancellationToken cancellationToken)
         {
             var voltageLevel = await _voltageLevelRepository.GetById(request.Id);
+            if (voltageLevel == null)
+                throw new ArgumentNotFoundException($"Сущность с заданным Id: {request.Id} не найдена.");
+
             await _voltageLevelRepository.Delete(voltageLevel);
 
             return Unit.Value;
